stitching/python/requirements.txt
glob2
numpy
Pillow
